Move integration tests
